Create README.md
I did not create the Tor Browser, but added a patch to it so you can run as root.
# Linux-Tor-Browser-patch
In contrast, I have added the patch to the #! tor-browser or you can add it for your self.
So, you can run it as root in Linux at your own risk!
Warranty Disclaimer: You use the code and suggestions at your own risk! 
Besides, you waive the right to tort claims and liabilities by downloading Using the code.
Considering it is at the sole discretion you waive the right to liability claims for its use or distribution.